+++ src/isolens/__init__.py
@@ -3,4 +3,4 @@
 A toolkit for analysis of long-read sequencing data.
 """
 
-__version__ = "0.5.2"
+__version__ = "0.5.3"